refactor(rspack_core): implement resolver proxy
This PR encapsulates all logic to `nodejs_resolver`,
and also adds all the placeholders for adding add oxc_resolver.
