+++ crates/rspack_core/Cargo.toml
@@ -21,7 +21,7 @@ hashlink = { workspace = true }
 indexmap = { workspace = true }
 itertools = { workspace = true }
 mime_guess = { workspace = true }
-nodejs-resolver = { workspace = true }
+nodejs-resolver = { version = "0.1.0" }
 once_cell = { workspace = true }
 paste = { workspace = true }
 petgraph = { version = "0.6.3", features = ["serde-1"] }
