+++ crates/rspack_core/src/lib.rs
@@ -76,6 +76,8 @@ mod ukey;
 pub use ukey::*;
 mod module_graph_module;
 pub use module_graph_module::*;
+pub mod resolver;
+pub use resolver::*;
 pub mod tree_shaking;
 
 pub use rspack_loader_runner::{get_scheme, ResourceData, Scheme, BUILTIN_LOADER_PREFIX};
